+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Added: experimental Gutenberg editor integration - [EnlighterJS/Plugin.Gutenberg on GitHub](https://github.com/EnlighterJS/Plugin.Gutenberg)
 * Added: keyboard shortcut `+x` to highlight inline code - requested by [dahnark on WordPress.org Forums](https://wordpress.org/support/topic/override-theme-code-style/)
 * Added: Visual Editor Plugin ([EnlighterJS.TinyMCE](https://github.com/EnlighterJS/Plugin.TinyMCE)) **v3.5.0**
+* Added: Environment check to ensure the plugin is uploaded into `wp-content/plugins/enlighter/`
 * Changed: moved experimental settings into panel "Beta"
 * Removed: outdated translations
 * Bugfix: styles of the "plugin upgrade notification" were broken